Multi tenant vhosts - Redefine Service URLs to be a function of instance.
All code which refers to a service URL needs to known which instance it is running as.
Work through the usages.

Switch service names to instance aware instance.domain vhost names.

auth, kahuna and thrall move to vhost url. Stops at extends Panda. We are going to replace Panda though.

loader, projection, usages, collections and leases move to vhost urls.
XXX Forces migration to be instance aware to project images!

cropper and metadata-editor are vhost based.

imgproxy is vhost based.

KahunaController will need to evaluate the service URLs to pass to the client on the fly.

This is a different concern to the config class which provides the function for doing this.

Catch all .rootUri usages in s strings in media-api.
Media API will probably render contextual links now.
s string substitutions allow alot of unevaluated functions to be rendered toString!