Fixes FileMetadataReaderTest fails locally during British summer by removing non time zoned date formatter which was shadowing the long standing time zoned date formatter.
Add a British summer time YYYY-MM-dd example to show that this formatter is locale dependant.

DateTimeFormat.forPattern("yyyy-MM-dd") matches the same pattern as ISODateTimeFormat.date.withZoneUTC but removes the withZoneUTC behaviour.

I do not know if that was intentional but FileMetadataReaderTest was a long standing test so this could be considered a regression.

Additionally, that entire block of date formatters probably have an indeterminate outcome.